[Change] separate books controllers from routes
